requirements.txt
flask
pymongo
python-dotenv
flask-cors
requests
fastapi
uvicorn
pandas
networkx
python-multipart
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc
grpcio
protobuf
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-pymongo